some aspect of testing

+++ app/src/main/java/com/example/user/finalproject/model/User_Information.java
@@ -4,7 +4,7 @@ package com.example.user.finalproject.model;
 public class User_Information {
     private String rowName;
     private String value;
-    private int db_ID;
+    private long db_ID;
 
     public User_Information(){
     }
@@ -25,11 +25,11 @@ public class User_Information {
         this.value = value;
     }
 
-    public int getDb_ID(){
+    public long getDb_ID(){
         return this.db_ID;
     }
 
-    public void setDb_ID(int db_ID){
+    public void setDb_ID(long db_ID){
         this.db_ID = db_ID;
     }
